Check for eof so we can exit gracefully
The TernaryTrees benchmark was reading from stdin until an execption
was raised. This caused it to exit with a non-zero exit code when
the end of input was reached. We now check for eof before reading
so we can exit with success. Any tools depending on exit code to
determine success should now be happy.