Updated benchmark source code for ghc > 7.0
The newer versions of ghc cannot work with both the haskell98 and
base packages. Any benchmarks with a cabal file that require
haskell98 package were changed to conditionally depend on haskell98
if impl(ghc) < 7.1.

The source code was updated so that imports used the fully qualified
name instead of the haskell98 name. This required the following
changes for imports:

  -import Array
  +import Data.Array

  -import Maybe
  -import Random
  +import Data.Maybe
  +import System.Random

  -import Char
  -import List
  +import Data.Char
  +import Data.List

  -import System
  +import System.Exit