Rename executables to match the benchmark name
This commit changes the executable name for the benchmark to be the
same as the name of the benchmark.

Previously the name of the benchmark executables did not match the
name of the benchmark.  This worked just fine, but when log files
were generated based on the name of the executable (as is the case
with DynamoRIO), it was sometimes a pain to map it back to the
actual benchmark. It is easier to process those log files if the
execuable name matches the benchmark name.