Add script to copy a helper source file to each benchmark
The script will copy the FibonHelper.hs file to each
benchmark so that it can be referenced in place rather than
a realative path to the location in the fibon repo. This
makes the benchmark repo able to stand on its own and also
allows for building benchmarks in parallel when dumping llvm
files. Since the llvm files are dumped next to the source
file the different benchmarks were clobbering eachother.